Windows build

+++ packages/seacas/libraries/ioss/src/Ioss_Utils.h
@@ -203,7 +203,7 @@ namespace Ioss {
       if (number == 0) {
         return 1;
       }
-      int width = std::floor(std::log10(number)) + 1;
+      int width = int(std::floor(std::log10(number))) + 1;
       if (use_commas) {
         width += ((width - 1) / 3);
       }
